Refactor permissions decorators and guard
This commit refactors the permissions decorators and guard in the auth module. It introduces a new decorator, CheckLoginStatus, which checks the login status of the user. It also modifies the ChurchRole decorator to use the userIdParam instead of the churchParam. Additionally, a new decorator, CheckChurch, is added to verify if the element being edited belongs to the authenticated user's church.

Related to #123