refactor: Update authentication utility functions
This commit updates the authentication utility functions in the `src/auth/utils` directory. The `checkAppRolesHandle` function now throws a `ForbiddenException` if the user does not have the required role. The `checkAdminHandle` function has been modified to check if the user is logged in and has the admin role. The `checkUserIdParamHandle` function now throws a `ForbiddenException` if the user ID in the request parameters does not match the session user ID. The `checkLoginStatusHandle` function now throws a `ForbiddenException` if the user is not logged in or is already logged in. These changes improve the security and functionality of the authentication system.